# swagger.yaml
openapi: 3.0.0
paths:
...
  /contacts:
    get:
      operationId: ContactsController_getContactsV1
      parameters:
        - name: search
          required: false
          in: query
          schema:
            type: string
        - name: nextPageToken
          required: false
          in: query
          schema:
            default: null
            type: string
        - name: limit
          required: false
          in: query
          schema:
            default: 50
            example: 50
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactsDto"
      security: &a5
        - bearer: []
      summary: Get list of contacts
      tags: &a6
        - Contacts
    post:
      operationId: ContactsController_createContactV1
      parameters: []
      requestBody:
        required: true
        description: Contact data
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactCreateDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactCreateResponseDto"
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactCreateResponseDto"
        "409":
          description: Contact already exists
        "500":
          description: Internal server error
      security: *a5
      summary: Create new contact
      tags: *a6
  /contacts/{id}:
    get:
      operationId: ContactsController_getContactV1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactDto"
      security: *a5
      summary: Get contact by ID
      tags: *a6
    put:
      operationId: ContactsController_updateContactV1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: string
      security: *a5
      summary: Update a contact
      tags: *a6
    delete:
      operationId: ContactsController_deleteContactV1
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: string
      security: *a5
      summary: Delete a contact
      tags: *a6
  /contacts-groups:
    get:
      operationId: ContactsGroupsControllerV1_getContactGroups
      parameters:
        - name: search
          required: false
          in: query
          schema:
            type: string
        - name: nextPageToken
          required: false
          in: query
          schema:
            default: null
            type: string
        - name: limit
          required: false
          in: query
          schema:
            default: 50
            example: 50
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactGroupsDto"
      security: &a7
        - bearer: []
      summary: Get list of contact's groups
      tags: &a8
        - Contacts
    post:
      operationId: ContactsGroupsControllerV1_getContactGroupV1
      parameters: []
      requestBody:
        required: true
        description: Contact Group DTO
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactGroupCreateDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactGroupDto"
      security: *a7
      summary: Create contact group
      tags: *a8
  /contacts-groups/{id}:
    get:
      operationId: ContactsGroupsControllerV1_getContactGroupByIdV1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactGroupDto"
        "404":
          description: Group requested not found
      security: *a7
      summary: Get contact groups
      tags: *a8
    put:
      operationId: ContactsGroupsControllerV1_updateAdministratorGroupV1
      parameters:
        - name: id
          required: true
          in: path
          description: Contact group ID
          schema:
            type: number
      requestBody:
        required: true
        description: Contact group DTO.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactGroupUpdateDto"
      responses:
        "200":
          description: Contact group updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactGroupDto"
        "406":
          description: Contact group could not be update.
      security: *a7
      summary: Update an contact group
      tags: *a8
    patch:
      operationId: ContactsGroupsControllerV1_patchContactGroupV1
      parameters:
        - name: id
          required: true
          in: path
          description: Contact group ID
          schema:
            type: number
      requestBody:
        required: true
        description: Patch operations.
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/PatchOperationDto"
      responses:
        "200":
          description: Contact group updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactGroupDto"
        "406":
          description: Contact group could not be update.
      security: *a7
      summary: Patch an contact group
      tags: *a8
  /pages:
    get:
      operationId: PagesControllerV1_getPagesV1
      parameters:
        - name: offset
          required: false
          in: query
          schema:
            default: 0
            example: 0
            type: number
        - name: limit
          required: false
          in: query
          schema:
            default: 50
            example: 50
            type: number
        - name: search
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagesListDto"
      security: &a9
        - bearer: []
      summary: Get list of pages
      tags: &a10
        - Pages
    post:
      operationId: PagesControllerV1_sendPagesV1
      parameters: []
      requestBody:
        required: true
        description: send Page DTO
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageSendDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageDto"
      security: *a9
      summary: Send page
      tags: *a10
  /pages/{id}:
    get:
      operationId: PagesControllerV1_getPageByIdV1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactDto"
      security: *a9
      summary: Get Page by ID
      tags: *a10
...
info:
  title: OnPage API
  description: >-
    OpenAPI definition for managing dispatching, pages, contacts,
    administrators, templates, settings, and webhooks.


    [📥 Download Swagger YAML](./swagger.yaml)
  version: "1.0"
  contact: {}
...
servers:
  - url: http://127.0.0.1:3003
    description: Development Server
  - url: https://public-api.onsetmobile.com
    description: Sandbox Server
  - url: https://public-api.onpage.com
    description: Production Server
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    Metadata:
      type: object
      properties:
        nextPageToken:
          type: string
      required:
        - nextPageToken
...
    ContactDto:
      type: object
      properties:
        id:
          type: number
          description: Account ID
        opid:
          type: string
          description: OnPage ID
        firstName:
          type: string
          description: Account first name
        lastName:
          type: string
          description: Account last name
        email:
          type: string
          description: Account email address
        phoneNumber:
          type: string
          description: Account phone number
        status:
          type: string
          description: Account OPID status
        groups:
          description: Groups the account is member of
          type: array
          items:
            type: string
      required:
        - id
        - opid
        - firstName
        - lastName
        - email
        - phoneNumber
        - status
    ContactsDto:
      type: object
      properties:
        contacts:
          description: Array of contacts
          type: array
          items:
            $ref: "#/components/schemas/ContactDto"
        metadata:
          description: Response metadata
          allOf:
            - $ref: "#/components/schemas/Metadata"
      required:
        - contacts
        - metadata
    ContactCreateDto:
      type: object
      properties:
        opid:
          type: string
          description: OnPage ID
        firstName:
          type: string
          description: Contact first name
        lastName:
          type: string
          description: Contact last name
        email:
          type: string
          description: Contact email address
        password:
          type: string
          description: Contact password
        phoneNumber:
          type: string
          description: Contact phone number
      required:
        - opid
        - firstName
        - lastName
        - email
        - password
        - phoneNumber
    IdResponseDto:
      type: object
      properties:
        id:
          type: number
          description: ID
      required:
        - id
    ContactCreateResponseDto:
      type: object
      properties:
        contact:
          description: Contact
          allOf:
            - $ref: "#/components/schemas/IdResponseDto"
      required:
        - contact
    ContactGroupMemberDto:
      type: object
      properties:
        contactId:
          type: number
          description: Contact ID
        order:
          type: number
          description: Escalation order
          nullable: true
      required:
        - contactId
    ESCALATION_INTERVAL_DTO:
      type: string
      enum:
        - NONE
        - 1 minute
        - 2 minutes
        - 3 minutes
        - 5 minutes
        - 10 minutes
        - 15 minutes
        - 20 minutes
        - 25 minutes
        - 30 minutes
        - 35 minutes
        - 40 minutes
        - 45 minutes
        - 50 minutes
        - 55 minutes
        - 1 hour
    ESCALATION_FACTOR:
      type: string
      enum:
        - NONE
        - DELIVERED
        - READ
        - REPLIED
    GroupFailOverDto:
      type: object
      properties:
        includeOriginalMessage:
          type: boolean
          description: Fail over includes origin al message
        emails:
          description: Fail over to emails list
          type: array
          items:
            type: string
        contacts:
          description: Fail over to contacts list
          type: array
          items:
            type: number
        groups:
          description: Fail over to contacts' groups list
          type: array
          items:
            type: number
      required:
        - includeOriginalMessage
        - emails
    ContactGroupDto:
      type: object
      properties:
        id:
          type: number
          description: Group ID
        opid:
          type: string
          description: OnPage ID (OPID)
        name:
          type: string
          description: Group name
        description:
          type: string
          description: Group description
        contacts:
          description: Contacts of the group
          type: array
          items:
            $ref: "#/components/schemas/ContactGroupMemberDto"
        escalation:
          type: boolean
          description: Escalation flag
        escalationInterval:
          description: Escalation interval
          allOf:
            - $ref: "#/components/schemas/ESCALATION_INTERVAL_DTO"
        escalationFactor:
          description: Escalation factor
          allOf:
            - $ref: "#/components/schemas/ESCALATION_FACTOR"
        failOver:
          description: Fail over parameters
          allOf:
            - $ref: "#/components/schemas/GroupFailOverDto"
      required:
        - id
        - opid
        - name
        - contacts
        - escalation
    ContactGroupsDto:
      type: object
      properties:
        contacts:
          description: Array of groups
          type: array
          items:
            $ref: "#/components/schemas/ContactGroupDto"
        metadata:
          description: Response metadata
          allOf:
            - $ref: "#/components/schemas/Metadata"
      required:
        - contacts
        - metadata
    ContactGroupCreateDto:
      type: object
      properties:
        opid:
          type: string
          description: OnPage ID (OPID)
        name:
          type: string
          description: Group name
        description:
          type: string
          description: Group description
        contacts:
          description: Contacts of the group
          type: array
          items:
            $ref: "#/components/schemas/ContactGroupMemberDto"
        escalation:
          type: boolean
          description: Escalation flag
        escalationInterval:
          description: Escalation interval
          allOf:
            - $ref: "#/components/schemas/ESCALATION_INTERVAL_DTO"
        escalationFactor:
          description: Escalation factor
          allOf:
            - $ref: "#/components/schemas/ESCALATION_FACTOR"
        failOver:
          description: Fail over parameters
          allOf:
            - $ref: "#/components/schemas/GroupFailOverDto"
      required:
        - opid
        - name
        - contacts
        - escalation
    ContactGroupUpdateDto:
      type: object
      properties:
        opid:
          type: string
          description: OnPage ID (OPID)
        name:
          type: string
          description: Group name
        description:
          type: string
          description: Group description
        contacts:
          description: Contacts of the group
          type: array
          items:
            $ref: "#/components/schemas/ContactGroupMemberDto"
        escalation:
          type: boolean
          description: Escalation flag
        escalationInterval:
          description: Escalation interval
          allOf:
            - $ref: "#/components/schemas/ESCALATION_INTERVAL_DTO"
        escalationFactor:
          description: Escalation factor
          allOf:
            - $ref: "#/components/schemas/ESCALATION_FACTOR"
        failOver:
          description: Fail over parameters
          allOf:
            - $ref: "#/components/schemas/GroupFailOverDto"
      required:
        - opid
        - name
        - contacts
        - escalation
    PatchOperationDto:
      type: object
      properties:
        op:
          type: string
          description: The operation to be performed
          enum:
            - add
            - remove
            - replace
            - move
            - copy
            - test
        path:
          type: string
          description: A JSON Pointer that indicates the location to perform the operation
        value:
          type: object
          description: The value to be used for the operation (if applicable)
        from:
          type: string
          description: The source location for move or copy operations (if applicable)
      required:
        - op
        - path
    PageFromDto:
      type: object
      properties:
        type:
          type: string
          description: Type
          enum:
            - EMAIL
            - DISPATCHER
            - ONPAGE
            - WEB
            - ACCOUNT
            - API
            - PHONE
            - WCTP
            - SNPP
        caption:
          type: string
          description: Caption
        value:
          type: object
          description: Value
      required:
        - type
        - caption
        - value
    PagePriority:
      type: string
      enum:
        - HIGH
        - LOW
    RecipientDto:
      type: object
      properties:
        type:
          type: string
          description: Type
          enum:
            - dispatcher
            - email
            - opid
            - system
        caption:
          type: string
          description: Caption
        value:
          type: string
          description: Type
      required:
        - type
        - caption
        - value
    PageDto:
      type: object
      properties:
        id:
          type: number
          description: Page ID
        chatId:
          type: number
          description: Chat ID
        from:
          description: Sender
          allOf:
            - $ref: "#/components/schemas/PageFromDto"
        subject:
          type: string
          description: Subject
        body:
          type: string
          description: Body
        priority:
          description: Priority
          default: HIGH
          allOf:
            - $ref: "#/components/schemas/PagePriority"
        recipients:
          description: Recipients
          type: array
          items:
            $ref: "#/components/schemas/RecipientDto"
        created:
          type: string
          description: Created at
        replies:
          description: Possible replies
          type: array
          items:
            type: string
        attachments:
          description: Attachments
          type: array
          items:
            type: string
        callbackUri:
          type: string
          description: Callback URI
      required:
        - id
        - chatId
        - from
        - subject
        - priority
        - recipients
        - created
    PagesListDto:
      type: object
      properties:
        pages:
          description: Array of pages
          type: array
          items:
            $ref: "#/components/schemas/PageDto"
        metadata:
          description: Response metadata
          allOf:
            - $ref: "#/components/schemas/Metadata"
      required:
        - pages
        - metadata
    PageSendDto:
      type: object
      properties:
        chatId:
          type: number
          description: Chat ID
        subject:
          type: string
          description: Subject
        body:
          type: string
          description: Body
        priority:
          description: Priority
          default: HIGH
          allOf:
            - $ref: "#/components/schemas/PagePriority"
        recipients:
          description: Recipients' OPIDs
          type: array
          items:
            type: string
        replies:
          description: Possible replies
          type: array
          items:
            type: string
        attachments:
          description: Attachments ID's list
          type: array
          items:
            type: string
        callbackUri:
          type: string
          description: Callback URI
      required:
        - chatId
        - subject
        - priority
        - recipients
...